@@ -1,7 +1,8 @@
 from fastapi import FastAPI
-from .api_book_management import app as book_management # relative import
-from .api_user_management import app as user_management # relative import
-from .api_borrow_return import app as borrow_return # relative import
+from .api_book_management import app as book_management 
+from .api_user_management import app as user_management 
+from .api_borrow_return import app as borrow_return 
+from .api_admin import app as admin 
 
 app = FastAPI(
     title="Library Management System",
@@ -11,8 +12,12 @@ app = FastAPI(
 
 app.mount("/book", book_management)
 app.mount("/user", user_management)
-app.mount("/borrows", borrow_return)
+app.mount("/borrow", borrow_return)
+app.mount("/admin", admin)
 
 @app.get("/", response_model=dict, status_code=200)
 async def root():
-    return {"title": app.title, "version": app.version, "description": app.description}
+    return {"title": app.title, 
+            "version": app.version, 
+            "description": app.description
+        }
